refactor: move ts types to their own folder
